At 02211455

+++ Foobar/Extentions/UIViewExt.swift
@@ -102,33 +102,33 @@ extension UIView {
 
   // 设置阴影
   func extSetShadow(color: UIColor = .black,
-                    opacity: Double = 0.5,
-                    radius: Double = 5.0,
+                    opacity: Float = 0.5,
+                    radius: CGFloat = 5.0,
                     offset: CGSize = .zero)
   {
     layer.masksToBounds = false
 
     layer.shadowColor = color.cgColor
-    layer.shadowOpacity = Float(opacity)
-    layer.shadowRadius = CGFloat(radius)
+    layer.shadowOpacity = opacity
+    layer.shadowRadius = radius
     layer.shadowOffset = offset
   }
 
   // 设置边框
-  func extSetBorder(width: Double = 1.0, color: UIColor = .lightGray) {
-    layer.borderWidth = CGFloat(width)
+  func extSetBorder(width: CGFloat = 1.0, color: UIColor = .lightGray) {
+    layer.borderWidth = width
     layer.borderColor = color.cgColor
   }
 
   // 设置圆角
-  func extSetCorner(radius: Double = 4.0) {
+  func extSetCorner(radius: CGFloat = 4.0) {
     layer.masksToBounds = true
-    layer.cornerRadius = CGFloat(radius)
+    layer.cornerRadius = radius
   }
 
 
   // 遮罩圆角, 如果视图设置了阴影, 阴影会被切掉
-  func extMaskRoundedLayer(radius: Double = 4.0, corners: UIRectCorner = .allCorners) {
+  func extMaskRoundedLayer(radius: CGFloat = 4.0, corners: UIRectCorner = .allCorners) {
     let shapeLayer = CAShapeLayer()
     let bezierPath = UIBezierPath(roundedRect: bounds,
                                   byRoundingCorners: corners,
@@ -142,9 +142,9 @@ extension UIView {
 
   // 添加圆角/有边框/无填充图层, 如果视图设置了阴影, 阴影会绕着这个图层的边界, 由于图层无填充, 边框内外都有阴影
   // 添加圆角/无边框/有填充图层, 如果视图设置了阴影, 阴影会绕着这个图层的边界
-  func extAddRoundedLayer(radius: Double = 4.0,
+  func extAddRoundedLayer(radius: CGFloat = 4.0,
                           corners: UIRectCorner = .allCorners,
-                          borderWidth: Double = 1.0,
+                          borderWidth: CGFloat = 1.0,
                           borderColor: UIColor = .lightGray,
                           fillColor: UIColor = .white)
   {
@@ -157,7 +157,7 @@ extension UIView {
                                   byRoundingCorners: corners,
                                   cornerRadii: CGSize(width: radius, height: radius))
     roundedLayer.path = bezierPath.cgPath
-    roundedLayer.lineWidth = CGFloat(borderWidth)
+    roundedLayer.lineWidth = borderWidth
     roundedLayer.strokeColor = borderColor.cgColor
     roundedLayer.fillColor = fillColor.cgColor
